+++ midilib/src/main/cpp/oscillators/ReverseSawtoothOscillator.cpp
@@ -1,6 +1,6 @@
 #include "ReverseSawtoothOscillator.h"
 
-float ReverseSawtoothOscillator::calculate(float time, float frequency, int8_t overtoneFactor) {
+float ReverseSawtoothOscillator::calculate(double time, float frequency, int8_t overtoneFactor) {
     float p = calcPhase(time, frequency, overtoneFactor);
     return mAmplitude * float(-p / M_PI);
 }
